+++ CHANGELOG.md
@@ -66,6 +66,8 @@
 - `tarantool::msgpack::skip_value` parameter trait bound has been mitigated, now
   accepts `&mut impl std::io::Read` instead of `&mut (impl std::io::Read + std::io::Seek)`
 - datetime `from_ffi_dt` and `as_ffi_dt` functions now public
+- `DecodeError` has now better error messages with report of deeply nested values
+  and actual MessagePack type is reported at type mismatch.
 
 ### Fixed